Refactored tests structure to use modules test_integration and test_unit
