--- src/codeschool/tests/test_codeschool.py
@@ -1,6 +1,0 @@
-import pytest
-import codeschool
-
-
-def test_some_basic_functionality():
-    assert 1 + 1 == 2
